Use bigint for iid

+++ db/migrate/20191127175335_add_iid_to_operations_feature_flags.rb
@@ -6,7 +6,7 @@ class AddIidToOperationsFeatureFlags < ActiveRecord::Migration[5.2]
   DOWNTIME = false
 
   def up
-    add_column :operations_feature_flags, :iid, :integer
+    add_column :operations_feature_flags, :iid, :bigint
   end
 
   def down
